+++ rector.php
@@ -19,24 +19,23 @@ return static function (RectorConfig $rectorConfig): void {
     ]);
 
     $rectorConfig->sets([
-        LevelSetList::UP_TO_PHP_72,
+        LevelSetList::UP_TO_PHP_83,
         SetList::CODE_QUALITY,
-        //SetList::DEAD_CODE,
         //SetList::PRIVATIZATION,
         //SetList::NAMING,
         //SetList::TYPE_DECLARATION,
         //SetList::EARLY_RETURN,
         //SetList::TYPE_DECLARATION_STRICT,
         SetList::DEAD_CODE,
-        PHPUnitSetList::PHPUNIT_CODE_QUALITY,
-        //PHPUnitSetList::PHPUNIT_90,
+        // PHPUnitSetList::PHPUNIT_CODE_QUALITY,
+        PHPUnitSetList::PHPUNIT_90,
         SetList::CODING_STYLE,
     ]);
 
     $rectorConfig->skip([
         StaticClosureRector::class,
         SimplifyIfElseToTernaryRector::class,
-        AddSeeTestAnnotationRector::class,
+        // AddSeeTestAnnotationRector::class,
     ]);
 
     $rectorConfig->importNames();